@@ -1,6 +1,6 @@
 class GroupsController < ApplicationController
   before_action :set_group, only: [:show, :update, :destroy, :small_groups, :grouping, :class_settings, :gender_setter]
-  before_filter :authenticate_user!
+  before_action :authenticate_user!
 
   def index
   end
@@ -67,7 +67,7 @@ class GroupsController < ApplicationController
   def class_settings
   end
 
-  private 
+  private
 
     def set_group
       @group = Group.find(params[:id])
